# MVP/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: World of Shadows — Backend API
  version: 1.0.0
  description: >-
    Bounded machine-readable API schema for the Flask backend /api/v1 HTTP surface.
    This artifact is authoritative for route inventory and OpenAPI-tag grouping inside
    this MVP package, but it is not the sole upstream authority for all overlapping
    runtime semantics. The higher-order runtime/API authority for play-service run
    lifecycle envelopes remains docs/technical/architecture/canonical_runtime_contract.md,
    and world-engine runtime authority remains governed by docs/technical/runtime/runtime-authority-and-state-flow.md
    and docs/ADR/adr-0001-runtime-authority-in-world-engine.md.
  x-wos-governance:
    artifact_role: canonical_api_schema
    authority_scope: bounded_backend_api_v1_route_inventory
    upstream_contracts:
      - docs/technical/architecture/canonical_runtime_contract.md
      - docs/technical/runtime/runtime-authority-and-state-flow.md
      - docs/ADR/adr-0001-runtime-authority-in-world-engine.md
    not_authoritative_for:
      - world-engine fastapi surface outside documented backend projection
      - websocket message contracts
      - runtime committed truth
      - unpublished or operator-only convenience notes
servers:
- url: /
  description: Same origin as backend (set base URL in your client).
...
paths:
...
  /api/v1/auth/forgot-password:
    post:
      tags:
      - Auth
      summary: POST /auth/forgot-password
      operationId: post_api_v1_auth_forgot_password
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      security: []
  /api/v1/auth/login:
    post:
      tags:
      - Auth
      summary: POST /auth/login
      operationId: post_api_v1_auth_login
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      security: []
  /api/v1/auth/logout:
    post:
      tags:
      - Auth
      summary: POST /auth/logout
      operationId: post_api_v1_auth_logout
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/auth/me:
    get:
      tags:
      - Auth
      summary: GET /auth/me
      operationId: get_api_v1_auth_me
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/auth/refresh:
    post:
      tags:
      - Auth
      summary: POST /auth/refresh
      operationId: post_api_v1_auth_refresh
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/auth/register:
    post:
      tags:
      - Auth
      summary: POST /auth/register
      operationId: post_api_v1_auth_register
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      security: []
  /api/v1/auth/resend-verification:
    post:
      tags:
      - Auth
      summary: POST /auth/resend-verification
      operationId: post_api_v1_auth_resend_verification
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
  /api/v1/auth/reset-password:
    post:
      tags:
      - Auth
      summary: POST /auth/reset-password
      operationId: post_api_v1_auth_reset_password
      responses:
        '200':
          description: Success (see docs/api/REFERENCE.md for response shapes).
        '401':
          description: Unauthorized — missing or invalid JWT where required.
        '403':
          description: Forbidden — policy or role.
        '404':
          description: Not found.
        '429':
          description: Rate limited.
        '500':
          description: Server error.
      security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Flask-JWT-Extended access token from `POST /api/v1/auth/login`.
security:
- BearerAuth: []